Change wording: compile -> index
Tom prefers this and this is what we are doing essentially.

I had other plans of "compiling" the templates down to something else.
Something like creating a META-template from templates. The idea was
that as the templates are being rendered, we will know how templates
are combined and we can create a META-template from them and
essentially make it a simple template with only string substitution
instead of multiple renders / joins cycle.

This seems to be too complex.